feat(v3design): adds v3_design flag
