+++ .env.testnet.local
@@ -53,3 +53,6 @@ NEXT_PUBLIC_PINATA_GATEWAY_KEY=
 
 # Reown
 NEXT_PUBLIC_REOWN_PROJECT_ID=1d97c3632e7aa07c9ca422c71125184a
+
+#Features
+NEXT_PUBLIC_ENABLE_FEATURE_V3_DESIGN=true
