+++ src/app/backing/components/Container/ActionMetricsContainer.tsx
@@ -12,10 +12,7 @@ export const ActionMetricsContainer: FC<ActionMetricsContainerProps> = ({
   return (
     <div
       data-testid={`${dataTestid}_ActionsMetrics`}
-      className={cn(
-        'flex flex-col w-[71.5rem] items-start gap-2 rounded-sm p-6 bg-v3-bg-accent-80',
-        className,
-      )}
+      className={cn('flex flex-col w-full items-start gap-2 rounded-sm p-6 bg-v3-bg-accent-80', className)}
     >
       <div className="flex items-start gap-14">{children}</div>
     </div>
