+++ src/app/backing/components/Container/ActionsContainer.tsx
@@ -16,11 +16,11 @@ export const ActionsContainer: FC<ActionsContainerProps> = ({
   return (
     <div
       data-testid={`${dataTestid}_ActionsContainer`}
-      className={cn('flex flex-col w-[71.5rem] gap-10 p-6 rounded-sm bg-v3-bg-accent-80', className)}
+      className={cn('flex flex-col w-full gap-10 p-6 rounded-sm bg-v3-bg-accent-80', className)}
     >
       <div
         data-testid={`${dataTestid}_ActionsContainerTitle`}
-        className="flex w-[66.5rem] justify-between items-center"
+        className="flex w-full justify-between items-center"
       >
         {title}
       </div>
